// src/03-widgets/articles/api/strategies/Article.abstract.ts
import { type Article } from '../types/Article'

export type ArticleResponse = {
  data: { articles: Article[] }
  error: null
} | {
  data: null
  error: unknown
}

export type FetchOptions = {
  page?: number
  pageSize?: number
  query?: string
}

export abstract class ArticleStrategy {
  abstract fetch(options: FetchOptions): Promise<ArticleResponse>
}